Visual Identity for Photon
For the future use of photon, it needs better recognition. (Not only the technology counts, but also the visual radiation).

We (myself and Thomas Ludwig) have designed a new brand, which fits better with Photon. This can be used for the future surface of wallets. Likewise be used in social media / exchanges / journalism.